requirements.txt
tqdm
numpy
scipy
pandas
pillow
matplotlib
scikit-learn
torch
torchvision
